Adjusting SSL Portion of HealthCheck
I ran into a issue running dashy in a podman oci container with the default value for public & private key paths  not being set the same way in healthcheck.js vs ssl-server.js

This difference was allowing SSL to start causing the healthcheck to get a 302 redirect to https .

This change adds in the same default path vars and i am going to say similar behavior that ssl-server.js uses. i tried to match the same style as the existing code between the two files

another alternative is to set `SSL_PUB_KEY_PATH` and `SSL_PRIV_KEY_PATH` in the compose file as env vars 

but after seeing #843 which lead me #768 and then finding  
#1410

ill be honest this probably solves #843 and #1410  i dont think this totally solved #768  tbh.  i had a hard time following that thread